# openapi.yaml
components:
...
    securitySchemes:
        anthropicGatewayKey:
            description: Gateway token on Anthropic-native ingress paths.
            in: header
            name: x-api-key
            type: apiKey
        bearerAuth:
            bearerFormat: API key
            description: 'Send Authorization: Bearer <token> when auth is enabled.'
            scheme: bearer
            type: http
        geminiGatewayKey:
            description: Gateway token on Gemini-native ingress paths.
            in: header
            name: x-goog-api-key
            type: apiKey
info:
    description: Compact provider-native LLM gateway with provider-aware routing, capability filtering, quota enforcement, SSE streaming, and provider-specific passthrough fields.
    title: llmgw
    version: v1
openapi: 3.0.3
paths:
...
    /v1/limits:
        get:
            operationId: getLimits
            responses:
                "200":
                    content:
                        application/json:
                            example:
                                key_id: demo-key
                                limits:
                                    daily_tokens: 1000000
                                    rpm: 60
                                source: dynamic
                                usage:
                                    daily_used_tokens: 240
                                    rpm_current: 3
                            schema:
                                properties:
                                    key_id:
                                        type: string
                                    limits:
                                        properties:
                                            budget_duration:
                                                type: string
                                            daily_tokens:
                                                format: int64
                                                type: integer
                                            max_input_tokens:
                                                format: int64
                                                type: integer
                                            max_output_tokens:
                                                format: int64
                                                type: integer
                                            max_parallel:
                                                format: int64
                                                type: integer
                                            max_spend_micros:
                                                format: int64
                                                type: integer
                                            model_allowlist:
                                                items:
                                                    type: string
                                                type: array
                                            monthly_tokens:
                                                format: int64
                                                type: integer
                                            provider_allowlist:
                                                items:
                                                    type: string
                                                type: array
                                            rpm:
                                                format: int64
                                                type: integer
                                            soft_spend_micros:
                                                format: int64
                                                type: integer
                                            tpm:
                                                format: int64
                                                type: integer
                                        type: object
                                    source:
                                        type: string
                                    usage:
                                        properties:
                                            active_requests:
                                                format: int64
                                                type: integer
                                            daily_held_tokens:
                                                format: int64
                                                type: integer
                                            daily_used_tokens:
                                                format: int64
                                                type: integer
                                            monthly_held_tokens:
                                                format: int64
                                                type: integer
                                            monthly_used_tokens:
                                                format: int64
                                                type: integer
                                            rpm_current:
                                                format: int64
                                                type: integer
                                            spend_held_micros:
                                                format: int64
                                                type: integer
                                            spend_used_micros:
                                                format: int64
                                                type: integer
                                            tpm_current:
                                                format: int64
                                                type: integer
                                        type: object
                                    usage_unavailable:
                                        type: boolean
                                type: object
                    description: Resolved quota limits and current usage for the authenticated key.
                "404":
                    content:
                        application/json:
                            example:
                                error:
                                    code: quota_limits_not_found
                                    message: no quota limits configured for this token
                                    type: invalid_request_error
                            schema:
                                properties:
                                    error:
                                        properties:
                                            code:
                                                type: string
                                            message:
                                                type: string
                                            param:
                                                type: string
                                            type:
                                                type: string
                                        type: object
                                type: object
                    description: No quota limits are configured for the authenticated key.
                default:
                    content:
                        application/json:
                            example:
                                error:
                                    message: unauthorized
                                    type: invalid_request_error
                            schema:
                                properties:
                                    error:
                                        properties:
                                            code:
                                                type: string
                                            message:
                                                type: string
                                            param:
                                                type: string
                                            type:
                                                type: string
                                        type: object
                                type: object
                    description: Error response.
            security:
                - bearerAuth: []
            summary: Get quota limits for the authenticated token key
            tags:
                - quota
        put:
            description: Requires the manage_limits principal permission.
            operationId: putLimits
            requestBody:
                content:
                    application/json:
                        example:
                            daily_tokens: 2000000
                            max_parallel: 8
                            max_spend_micros: 5000000
                            rpm: 120
                            tpm: 500000
                        schema:
                            properties:
                                budget_duration:
                                    type: string
                                daily_tokens:
                                    format: int64
                                    type: integer
                                max_input_tokens:
                                    format: int64
                                    type: integer
                                max_output_tokens:
                                    format: int64
                                    type: integer
                                max_parallel:
                                    format: int64
                                    type: integer
                                max_spend_micros:
                                    format: int64
                                    type: integer
                                model_allowlist:
                                    items:
                                        type: string
                                    type: array
                                monthly_tokens:
                                    format: int64
                                    type: integer
                                provider_allowlist:
                                    items:
                                        type: string
                                    type: array
                                rpm:
                                    format: int64
                                    type: integer
                                soft_spend_micros:
                                    format: int64
                                    type: integer
                                tpm:
                                    format: int64
                                    type: integer
                            type: object
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            example:
                                key_id: demo-key
                                limits:
                                    daily_tokens: 2000000
                                    rpm: 120
                                source: dynamic
                            schema:
                                properties:
                                    key_id:
                                        type: string
                                    limits:
                                        properties:
                                            budget_duration:
                                                type: string
                                            daily_tokens:
                                                format: int64
                                                type: integer
                                            max_input_tokens:
                                                format: int64
                                                type: integer
                                            max_output_tokens:
                                                format: int64
                                                type: integer
                                            max_parallel:
                                                format: int64
                                                type: integer
                                            max_spend_micros:
                                                format: int64
                                                type: integer
                                            model_allowlist:
                                                items:
                                                    type: string
                                                type: array
                                            monthly_tokens:
                                                format: int64
                                                type: integer
                                            provider_allowlist:
                                                items:
                                                    type: string
                                                type: array
                                            rpm:
                                                format: int64
                                                type: integer
                                            soft_spend_micros:
                                                format: int64
                                                type: integer
                                            tpm:
                                                format: int64
                                                type: integer
                                        type: object
                                    source:
                                        type: string
                                    usage:
                                        properties:
                                            active_requests:
                                                format: int64
                                                type: integer
                                            daily_held_tokens:
                                                format: int64
                                                type: integer
                                            daily_used_tokens:
                                                format: int64
                                                type: integer
                                            monthly_held_tokens:
                                                format: int64
                                                type: integer
                                            monthly_used_tokens:
                                                format: int64
                                                type: integer
                                            rpm_current:
                                                format: int64
                                                type: integer
                                            spend_held_micros:
                                                format: int64
                                                type: integer
                                            spend_used_micros:
                                                format: int64
                                                type: integer
                                            tpm_current:
                                                format: int64
                                                type: integer
                                        type: object
                                    usage_unavailable:
                                        type: boolean
                                type: object
                    description: Stored quota limits for the authenticated key.
                default:
                    content:
                        application/json:
                            example:
                                error:
                                    message: invalid request
                                    type: invalid_request_error
                            schema:
                                properties:
                                    error:
                                        properties:
                                            code:
                                                type: string
                                            message:
                                                type: string
                                            param:
                                                type: string
                                            type:
                                                type: string
                                        type: object
                                type: object
                    description: Error response.
            security:
                - bearerAuth: []
            summary: Set quota limits for the authenticated token key
            tags:
                - quota
...
servers:
    - description: Configured gateway listener
      url: /
...